Refactor simulation logic into Backbone model
